Do not relay on vboxmanage output.
There is a function which retrieve information about VM by looking on
the 'vboxmanage vminfo' command. The output wasn't consistent, yet it
was fixed recently so that amount of memory was all wrong.

Currently the only thing which is searched for is the VM configuration
path, and all the rest is taken out from the XML VM definition.